engine/pyproject.toml
[project]
name = "email-digest"
version = "2.0.0"
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.28.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.120.0",
    "anthropic>=0.25.0",
    "httpx>=0.27.0",
    "supabase>=2.4.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "fpdf2>=2.7.0",
    "python-docx>=1.1.0",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".."]
asyncio_mode = "auto"
